requirements.txt
pywin32==306
python-dotenv==1.0.1
pandas==2.2.0
openpyxl==3.1.2
pyside6==6.6.1
ollama==0.1.5
python-dateutil==2.8.2